Add confirmation when deleting

+++ src/app/user/user.component.ts
@@ -4,6 +4,7 @@ import { UserService, User } from '../core/user.service';
 import { MatDialog } from '@angular/material';
 import { UserEditDialogComponent } from './user-edit-dialog/user-edit-dialog.component';
 import { UserCreateDialogComponent } from './user-create-dialog/user-create-dialog.component';
+import { DialogService } from '../shared/dialog/dialog.service';
 
 @Component({
   selector: 'app-user',
@@ -17,7 +18,8 @@ export class UserComponent implements OnInit {
 
   constructor(
     private userService: UserService,
-    private dialog: MatDialog
+    private dialog: MatDialog,
+    private dialogService: DialogService
   ) { }
 
   ngOnInit() {
@@ -44,10 +46,17 @@ export class UserComponent implements OnInit {
     });    
   } 
 
-  public delete(user: User) {
-    this.userService.deleteUser(user).subscribe(x => {
-      this.ngOnInit()
+  public delete(user: User) {    
+    let message = `Are you sure you want to delete the user <em>${user.firstName} ${user.lastName} (${user.code})</em> ?`;
+    this.dialogService.showConfirmation(message).subscribe(x =>{ 
+      if(!x)
+        return;
+
+        this.userService.deleteUser(user).subscribe(x => {
+          this.ngOnInit()
+        });
     });
+    
   }
 }
